requirements.txt
asyncio
pyrogram
tgcrypto
python-dotenv
pymongo
